+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
-## v49.x / v50 (In progress)
+## v50 (2022-02-11)
 - Ancient Cave: Jeff can fix any item at IQ 1.
 - Holiday: Christmas and Valentine's Day options added.
+- NPC sprites: Added new flag level 2, random characters includes custom sprites.
 - PC sprites: Added "Random any" as new default, improved Random vanilla.
 - PC sprites: Added Exit Mouse, Rope Snake, Coil Snake, Duster, Flint, Save Frog, Edgar, Rek, General Leo, Relm, Strago, Prince of Cannock, Prince of Midenhall, Princess of Moonbrooke, Pit, Dark Pit, Mike, Dorothy, Sophia, Blanche, and Rose.
 - Music: Some custom tracks added.